Extract MagicNamespace from Templater
This is getting quite large, and I also think I'll need to create a
`MagicNamespace` instance outside of the `Templater`, therefore it makes
sense to extract this now.
